pyproject.toml
[tool.poetry]
name = "project_catalog"
version = "0.1.0"
description = "A CNN-based document processing and classification system"
authors = ["Callum Doty <doty.callum9@gmail.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
dropbox = "^11.36.2"
torch = "^2.1.1"
torchvision = "^0.16.1"
pillow = "^10.1.0"
redis = "^5.0.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3